Entity 식별자 타입 변경 Integer -> Long // 게시판 엔티티 생성 Post, Comment
생성시간과 수정시간을 관리하기위한 BaseEntity 추가

+++ src/main/java/com/dfo/dunsee/member/repository/BookmarkRepository.java
@@ -5,7 +5,7 @@ import com.dfo.dunsee.member.entity.CharacterInfo;
 import com.dfo.dunsee.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookmarkRepository extends JpaRepository<Bookmark, Integer>, BookmarkCustomRepository {
+public interface BookmarkRepository extends JpaRepository<Bookmark, Long>, BookmarkCustomRepository {
 
   Bookmark findByMemberAndCharacterInfo(Member member, CharacterInfo characterInfo);
